refactor(src): 훅, 유틸 import/export 방식 변경 및 반영
