+++ frontend/src/pages/TemplateList.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 import { Flex, TemplateItem, Text } from '@/components';
-import { useTemplateListQuery } from '@/hooks/useTemplateListQuery';
+import { useTemplateListQuery } from '@/hooks/template';
 
 const TemplateList = () => {
   const { data, error, isLoading } = useTemplateListQuery();
