+++ frontend/src/utils/index.ts
@@ -1,0 +1,1 @@
+export { formatRelativeTime } from './formatRelativeTime';
